Created script to change org headline divs to sections for reveal.js.
I use org-special-blocks-extra for examples, lemmas, theorems etc. But ox-reveal does not seem to work with these blocks. So I export to html which does work with these blocks, and then I run the script to convert the divs into sections for reveal.js. Hack, yes. But it means my src doc is not polluted by extraneous tags and code so I can focus on writing the content.